# Conflicts:
#	tutorials/common/lights/light.cpp
#	tutorials/common/math/vec.h
#	tutorials/common/tutorial/optics.h
#	tutorials/common/tutorial/tutorial_device.ispc
#	tutorials/interpolation/interpolation_device.cpp
#	tutorials/ispc2cpp.sh

+++ common/sys/barrier.h
@@ -106,7 +106,7 @@ namespace embree
     volatile unsigned int mode;
     volatile unsigned int flag0;
     volatile unsigned int flag1;
-    volatile unsigned int threadCount;
+    volatile size_t threadCount;
   };
 }
