+++ kernels/geometry/object.h
@@ -54,8 +54,8 @@ namespace embree
     __forceinline std::pair<BBox3fa,BBox3fa> fill_mblur(const PrimRef* prims, size_t& i, size_t end, Scene* scene, const bool list)
     {
       const PrimRef& prim = prims[i]; i++;
-      const size_t geomID = prim.geomID();
-      const size_t primID = prim.primID();
+      const unsigned geomID = prim.geomID();
+      const unsigned primID = prim.primID();
       new (this) Object(geomID, primID);
       AccelSet* accel = (AccelSet*) scene->get(geomID);
       return accel->bounds_mblur(primID);
