+++ kernels/geometry/trianglev_mb.h
@@ -157,8 +157,8 @@ namespace embree
       for (size_t i=0; i<M && begin<end; i++, begin++)
       {
 	const PrimRef& prim = prims[begin];
-        const size_t geomID = prim.geomID();
-        const size_t primID = prim.primID();
+        const unsigned geomID = prim.geomID();
+        const unsigned primID = prim.primID();
         const TriangleMesh* __restrict__ const mesh = scene->getTriangleMesh(geomID);
         const TriangleMesh::Triangle& tri = mesh->triangle(primID);
 	const Vec3fa& a0 = mesh->vertex(tri.v[0],0); bounds0.extend(a0);
