+++ kernels/subdiv/half_edge.h
@@ -48,7 +48,7 @@ namespace embree
     }
     
     __forceinline bool hasOpposite() const { return opposite_half_edge_ofs != 0; }
-    __forceinline void setOpposite(HalfEdge* opposite) { opposite_half_edge_ofs = opposite-this; }
+    __forceinline void setOpposite(HalfEdge* opposite) { opposite_half_edge_ofs = int(opposite-this); }
     
     __forceinline       HalfEdge* next()       { assert( next_half_edge_ofs != 0 ); return &this[next_half_edge_ofs]; }
     __forceinline const HalfEdge* next() const { assert( next_half_edge_ofs != 0 ); return &this[next_half_edge_ofs]; }
@@ -194,9 +194,9 @@ namespace embree
     }
     
     /*! counts number of polygon edges  */
-    __forceinline size_t numEdges() const
+    __forceinline unsigned int numEdges() const
     {
-      size_t N = 1;
+      unsigned int N = 1;
       for (const HalfEdge* p=this->next(); p!=this; p=p->next(), N++);
       return N;
     }
